% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ansari_Bradley_test.R
\name{Ansari_Bradley_test}
\alias{Ansari_Bradley_test}
\title{Ansari-Bradley Test of Homogeneity of Variances}
\usage{
Ansari_Bradley_test(
  data,
  formula,
  alpha = 0.05,
  silent = FALSE,
  summary = FALSE,
  misc = FALSE
)
}
\arguments{
\item{data}{A data frame containing the variables specified in the formula.}

\item{formula}{A formula of the form \code{DV ~ IV}, where \code{DV} is the dependent
(response) variable and \code{IV} is the independent (grouping) variable.}

\item{alpha}{A numeric value specifying the significance level. Must be
between 0 and 1. Default is 0.05.}

\item{silent}{A logical value. If \code{FALSE} (default), results are
printed to the console. If \code{TRUE}, no output is printed.}

\item{summary}{A logical value (default: \code{FALSE}). If \code{TRUE}, a summary table
for the input data is returned.}

\item{misc}{A logical value. If \code{FALSE} (default), only essential
parameters are returned. If \code{TRUE}, additional auxiliary
parameters are included in the output.}
}
\value{
A list containing the test statistics, p-value, degrees of freedom,
and optionally a summary table and/or auxiliary parameters,
depending on the values of \code{summary} and \code{misc}.
}
\description{
Performs Ansari-Bradley test to assess the null hypothesis that the variances
are equal across all groups (samples) defined by the independent variable.
}
\details{
This function is based on Conover et al. (1981) and uses an approximate chi-square
test. Results may be inaccurate for small total sample sizes (N < 20) or when
there are many tied values.
}
\examples{
df0 <- CYCB1[[1]]
out <- Ansari_Bradley_test(df0, cells ~ grp)
boxplot(cells ~ grp, df0, horizontal = TRUE)
points(x = df0$cells, y = jitter(as.numeric(df0$grp), amount = 0.15))
}
\references{
Ansari, A. R., & Bradley, R. A. (1960).
Rank-sum tests for dispersions.
The Annals of Mathematical Statistics, 31, 1174–1189.

Conover, W. J., Johnson, M. E., & Johnson, M. M. (1981).
A comparative study of tests for homogeneity of variances, with applications
to the outer continental shelf bidding data.
Technometrics, 23, 351–361.
https://doi.org/10.1080/00401706.1981.10487680
}
\seealso{
\link[=Fligner_Killeen_test]{stats::ansari.test}
}
