% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYCB1}
\alias{CYCB1}
\title{Cell number}
\format{
A list containing three data frames, each representing a different experimental batch:
\describe{
\item{TEMP}{Air temperature in degrees Celsius}
\item{RGF1}{Concentration of RGF1 peptide hormone treatment (0 nM, 5 nM)}
\item{treatment}{TEMP x RGF1 -> 4 groups of treatments}
\item{grp}{Labels for each treatment groups (A, B, C, D)}
\item{cells}{Number of meristematic root cells}
}
}
\usage{
CYCB1
}
\description{
A dataset containing cortex cell counts within the meristematic zone of roots
in an \emph{Arabidopsis thaliana} transgenic plant (CYCB1;3-GFP).
}
\references{
Hsiao, Y., Lai, J., Shiue, S., & Yamada, M. (2026).
RGF signaling bridges root development and nonlethal thermal stress adaptation.
New Phytologist, nph.71392.
https://doi.org/10.1111/nph.71392
}
\keyword{datasets}
