% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fligner_Killeen_test.R
\name{Fligner_Killeen_test}
\alias{Fligner_Killeen_test}
\title{Fligner-Killeen Test of Homogeneity of Variances}
\usage{
Fligner_Killeen_test(
  data,
  formula,
  alpha = 0.05,
  silent = FALSE,
  summary = FALSE,
  misc = FALSE
)
}
\arguments{
\item{data}{A data frame containing the variables specified in the formula.}

\item{formula}{A formula of the form \code{DV ~ IV}, where \code{DV} is the dependent
(response) variable and \code{IV} is the independent (grouping) variable.}

\item{alpha}{A numeric value specifying the significance level. Must be
between 0 and 1. Default is 0.05.}

\item{silent}{A logical value. If \code{FALSE} (default), results are
printed to the console. If \code{TRUE}, no output is printed.}

\item{summary}{A logical value (default: \code{FALSE}). If \code{TRUE}, a summary table
for the input data is returned.}

\item{misc}{A logical value. If \code{FALSE} (default), only essential
parameters are returned. If \code{TRUE}, additional auxiliary
parameters are included in the output.}
}
\value{
A list containing the test statistics, p-value, degrees of freedom,
and optionally a summary table and/or auxiliary parameters,
depending on the values of \code{summary} and \code{misc}.
}
\description{
Performs Fligner-Killeen test to assess the null hypothesis that the variances
are equal across all groups (samples) defined by the independent variable.
}
\examples{
df0 <- CYCB1[[1]]
out <- Fligner_Killeen_test(df0, cells ~ grp)
boxplot(cells ~ grp, df0, horizontal = TRUE)
points(x = df0$cells, y = jitter(as.numeric(df0$grp), amount = 0.15))
}
\references{
Fligner, M. A., & Killeen, T. J. (1976).
Distribution-free two-sample tests for scale.
Journal of the American Statistical Association, 71(353), 210–213.
https://doi.org/10.1080/01621459.1976.10481517

Conover, W. J., Johnson, M. E., & Johnson, M. M. (1981).
A comparative study of tests for homogeneity of variances, with applications
to the outer continental shelf bidding data.
Technometrics, 23, 351–361.
https://doi.org/10.1080/00401706.1981.10487680
}
\seealso{
\link[=Ansari_Bradley_test]{stats::fligner.test}
}
