% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_var_equal.R
\name{is_var_equal}
\alias{is_var_equal}
\title{Homogeneity of variance test}
\usage{
is_var_equal(
  data,
  formula = NULL,
  alpha = 0.05,
  silent = TRUE,
  summary = FALSE,
  sensitivity = 3
)
}
\arguments{
\item{data}{A data frame or a list of numeric vectors.}

\item{formula}{Formula (default: NULL).
If \code{data} is a data frame, define the val ~ group.}

\item{alpha}{Significance threshold, range from 0 to 1 (default: 0.05).}

\item{silent}{A logical value. If \code{FALSE} (default), results are
printed to the console. If \code{TRUE}, no output is printed.}

\item{summary}{Logical (default: FALSE). If \code{TRUE}, show the summary table.}

\item{sensitivity}{Numeric, range from 1 to 5 (default: 3).
The greater the value, the greater chance to consider as variance not equal.}
}
\value{
A boolean value or a list if the \code{summary} is set to \code{TRUE}.
}
\description{
A convenient wrapper that returns a logical value indicating whether variances
across groups are equal. It makes its decision based on the results of several
homoscedasticity tests, including the modified Brown–Forsythe (MBF), Fligner–Killeen (FK),
't Lam's G (LG), Levene's (LV), and O'Neill–Mathews (OM) tests.
}
\examples{
is_var_equal(roGFP[[1]], ro ~ grp)
}
