% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O.Brien_test.R
\name{O.Brien_test}
\alias{O.Brien_test}
\title{O'Brien's Test of Homogeneity of Variances}
\usage{
O.Brien_test(
  data,
  formula,
  alpha = 0.05,
  silent = FALSE,
  summary = FALSE,
  misc = FALSE,
  transform = function(x) (x - stats::median(x))^2
)
}
\arguments{
\item{data}{A data frame containing the variables specified in the formula.}

\item{formula}{A formula of the form \code{DV ~ IV}, where \code{DV} is the dependent
(response) variable and \code{IV} is the independent (grouping) variable.}

\item{alpha}{A numeric value specifying the significance level. Must be
between 0 and 1. Default is 0.05.}

\item{silent}{A logical value. If \code{FALSE} (default), results are
printed to the console. If \code{TRUE}, no output is printed.}

\item{summary}{A logical value (default: \code{FALSE}). If \code{TRUE}, a summary table
for the input data is returned.}

\item{misc}{A logical value. If \code{FALSE} (default), only essential
parameters are returned. If \code{TRUE}, additional auxiliary
parameters are included in the output.}

\item{transform}{A function used to transform the response variable into
deviations from a specified location measure.}
}
\value{
A list containing the test statistics, p-value, degrees of freedom,
and optionally a summary table and/or auxiliary parameters,
depending on the values of \code{summary} and \code{misc}.
}
\description{
Performs O'Brien's test to assess the null hypothesis that the variances
are equal across all groups (samples) defined by the independent variable.
}
\details{
\code{transform}:
The concept is similar to ANOVA procedure (transform the response variable to
residuals before analysis). Possible transformation are:
\itemize{
\item y' = |yi - ybar|
\item y' = (yi - ybar) ^ 2  (default)
\item y' = ln((yi - ybar) ^ 2)
\item y' = sqrt(|yi - ybar|)
}

The \code{ybar} could be either mean, median (default), or trimmed-mean.

\emph{Note:} This test is often regarded as conservative and may have relatively
low power to detect heteroscedasticity. The Levene and Brown-Forsythe tests
are generally preferred for assessing the homogeneity of variances.
}
\examples{
df0 <- CYCB1[[1]]
out <- O.Brien_test(df0, cells ~ grp)
boxplot(cells ~ grp, df0, horizontal = TRUE)
points(x = df0$cells, y = jitter(as.numeric(df0$grp), amount = 0.15))
}
\references{
O’Brien, R. G. (1981).
A simple test for variance effects in experimental designs.
Psychological Bulletin, 89(3), 570–574.
https://doi.org/10.1037/0033-2909.89.3.570
}
\seealso{
[Brown_Forsythe_test][Levene_test][O.Neill_Mathews_test]
}
