% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O.Neill_Mathews_test.R
\name{O.Neill_Mathews_test}
\alias{O.Neill_Mathews_test}
\title{O'Neill-Mathews Test of Homogeneity of Variances}
\usage{
O.Neill_Mathews_test(
  data,
  formula,
  alpha = 0.05,
  silent = FALSE,
  summary = FALSE,
  misc = FALSE,
  transform = function(x) abs(x - stats::median(x))
)
}
\arguments{
\item{data}{A data frame containing the variables specified in the formula.}

\item{formula}{A formula of the form \code{DV ~ IV}, where \code{DV} is the dependent
(response) variable and \code{IV} is the independent (grouping) variable.}

\item{alpha}{A numeric value specifying the significance level. Must be
between 0 and 1. Default is 0.05.}

\item{silent}{A logical value. If \code{FALSE} (default), results are
printed to the console. If \code{TRUE}, no output is printed.}

\item{summary}{A logical value (default: \code{FALSE}). If \code{TRUE}, a summary table
for the input data is returned.}

\item{misc}{A logical value. If \code{FALSE} (default), only essential
parameters are returned. If \code{TRUE}, additional auxiliary
parameters are included in the output.}

\item{transform}{A function used to transform the response variable into
deviations from a specified location measure.}
}
\value{
A list containing the test statistics, p-value, degrees of freedom,
and optionally a summary table and/or auxiliary parameters,
depending on the values of \code{summary} and \code{misc}.
}
\description{
Performs O'Neill-Mathews test to assess the null hypothesis that the variances
are equal across all groups (samples) defined by the independent variable.
}
\details{
\code{transform}:
The concept is similar to ANOVA procedure (transform the response variable to
residuals before analysis). Possible transformation are:
\itemize{
\item y' = |yi - ybar| (default)
\item y' = (yi - ybar) ^ 2
\item y' = ln((yi - ybar) ^ 2)
\item y' = sqrt(|yi - ybar|)
}

The \code{ybar} could be either mean, median (default), or trimmed-mean.
}
\examples{
df0 <- roGFP[[1]]
out <- O.Neill_Mathews_test(df0, ro ~ grp)
boxplot(ro ~ grp, df0, horizontal = TRUE)
points(x = df0$ro, y = jitter(as.numeric(df0$grp), amount = 0.15))
}
\references{
O’Neill, M. E., & Mathews, K. (2000).
Theory & Methods: A Weighted Least Squares Approach to Levene’s Test of Homogeneity of Variance.
Australian & New Zealand Journal of Statistics, 42(1), 81–100.
https://doi.org/10.1111/1467-842X.00109
}
\seealso{
\link{Levene_test} \link{Brown_Forsythe_test}
}
