% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brown_Forsythe_test.R
\name{Brown_Forsythe_test}
\alias{Brown_Forsythe_test}
\title{Brown-Forsythe Test of Homogeneity of Variances}
\usage{
Brown_Forsythe_test(
  data,
  formula,
  alpha = 0.05,
  silent = FALSE,
  summary = TRUE,
  misc = FALSE
)
}
\arguments{
\item{data}{A data frame containing the variables specified in the formula.}

\item{formula}{A formula of the form \code{DV ~ IV}, where \code{DV} is the dependent
(response) variable and \code{IV} is the independent (grouping) variable.}

\item{alpha}{A numeric value specifying the significance level. Must be
between 0 and 1. Default is 0.05.}

\item{silent}{A logical value. If \code{FALSE} (default), results are
printed to the console. If \code{TRUE}, no output is printed.}

\item{summary}{A logical value. If \code{TRUE} (default), a summary table
of the test results is returned.}

\item{misc}{A logical value. If \code{FALSE} (default), only essential
parameters are returned. If \code{TRUE}, additional auxiliary
parameters are included in the output.}
}
\value{
A list containing the test statistics, p-value, degrees of freedom,
and optionally a summary table and/or auxiliary parameters,
depending on the values of \code{summary} and \code{misc}.
}
\description{
Performs Brown-Forsythe test to assess the null hypothesis that the variances
are equal across all groups (samples) defined by the independent variable. This
test is essentially Levene's test using medians from the various groups.
}
\examples{
plasma_etching <- data.frame(
    etch_rate = c(575, 542, 530, 539, 570,
                  565, 593, 590, 579, 610,
                  600, 651, 610, 637, 629,
                  725, 700, 715, 685, 710),
    power = rep(c(160, 180, 200, 220), each = 5)
)
Brown_Forsythe_test(plasma_etching, etch_rate ~ power)
}
\references{
Brown, M. B., & Forsythe, A. B. (1974).
Robust tests for the equality of variances.
Journal of the American Statistical Association, 69(346), 364–367.
https://doi.org/10.1080/01621459.1974.10482955
}
\seealso{
\code{\link[=Levene_test]{Levene_test()}}
}
