% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_var_equal.R
\name{check_var_equal}
\alias{check_var_equal}
\title{Homogeneity of variance test}
\usage{
check_var_equal(
  data,
  formula,
  alpha = 0.05,
  method = "LV",
  silent = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame or a list of numeric vectors.}

\item{formula}{Formula (default: NULL).
If \code{data} is a data frame, define the val ~ group.}

\item{alpha}{Significance threshold, range from 0 to 1 (default: 0.05).}

\item{method}{Character (default: "LV"). Abbreviation specifying the normality test to
perform. Available options are \code{c("AB", "BL", "FK", "LG", "LV", "MBF", "OB", "OM")}.}

\item{silent}{A logical value. If \code{FALSE} (default), results are
printed to the console. If \code{TRUE}, no output is printed.}

\item{...}{Additional arguments passed to the selected test function.}
}
\value{
A list.
}
\description{
A wrapper function for the homoscedasticity tests available in this package.
}
\details{
The \code{method} argument specifies the statistical procedure used to assess
whether group variances are equal.

Available methods are:
\itemize{
\item \code{"AB"}: Ansari-Bradley test
A rank-based nonparametric test for homogeneity of scale (dispersion) across
groups. It is often used as an alternative to the F-test when data are non-normal.
However, the Fligner–Killeen test is generally more robust and is recommended
for this purpose.
\item \code{"BL"}: Bartlett test
A classical parametric test for assessing homoscedasticity across multiple
groups. It is highly sensitive to outliers and deviations from normality.
When the data are normally distributed and free of outliers, it is the most
robust and powerful tests for equality of variances.
\item \code{"FK"}: Fligner–Killeen test
A rank-based nonparametric test for homogeneity of variances across groups.
It is based on absolute deviations from the median and is highly robust to
non-normality and outliers.
\item \code{"LG"}: 't Lam's G test
An extension of Cochran's C test used to evaluate the internal consistency of
variances. Although primarily designed as a variance outlier detection method
rather than a formal test of homoscedasticity (such as Levene's or Bartlett's test),
it can be used informally to assess variance homogeneity across groups. Its
robustness is particularly strong under normality and in the absence of outliers.
\item \code{"LV"}: Levene's test
A classical test for equality of variances based on an ANOVA framework.
This implementation uses the median as the center (instead of the mean),
making it more robust to non-normality and outliers.
\item \code{"MBF"}: Brown–Forsythe test (modified by Mehrotra)
A robust modification of Levene's test proposed by Brown and Forsythe.
The degrees of freedom are further adjusted following Mehrotra's correction,
improving performance under non-normality and in the presence of outliers.
\item \code{"OB"}: O'Brien test
A variance homogeneity test similar in spirit to Levene's test, based on
transformed observations designed to reduce sensitivity to non-normality.
\item \code{"OM"}: O'Neill–Mathews test
A modified Levene-type procedure that uses a weighted least squares approach.
}

In all methods, the null hypothesis is that group variances are equal.
}
\examples{
check_var_equal(roGFP[[1]], ro ~ grp, method = "LV")
}
