% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roGFP}
\alias{roGFP}
\title{RO index}
\format{
A list containing three data frames, each representing a different experimental batch:
\describe{
\item{TEMP}{Air temperature in degrees Celsius (22C and 31C)}
\item{RGF1}{Concentration of RGF1 peptide hormone treatment (0nM, 5nM)}
\item{treatment}{TEMP x RGF1 -> 4 groups of treatments}
\item{grp}{Handy labels for each treatment (A, B, C, D)}
\item{ro}{Redox index ranging from -1 (reduced) to 1 (oxidized)}
}
}
\usage{
roGFP
}
\description{
A dataset containing the reduced–oxidized (RO) index computed from confocal
fluorescence images of roots of an \emph{Arabidopsis thaliana} transgenic plant
(cytRGX–roGFP2).
}
\keyword{datasets}
